refactor enums plan
