+++ lib/models/enums/plan_enums.dart
@@ -1,0 +1,11 @@
+
+import 'package:flutter/material.dart';
+
+import '../../custom_icons.dart';
+
+part 'bike_rental_network.dart';
+part 'biking_speed.dart';
+part 'optimize_type.dart';
+part 'transport_mode.dart';
+part 'walk_board_cost.dart';
+part 'walking_speed.dart';
