+++ lib/models/enums/walk_board_cost.dart
@@ -1,0 +1,11 @@
+part of 'plan_enums.dart';
+
+enum WalkBoardCost { defaultCost, walkBoardCostHigh }
+
+extension WalkBoardCostExtension on WalkBoardCost {
+  static const values = <WalkBoardCost, int>{
+    WalkBoardCost.defaultCost: 600,
+    WalkBoardCost.walkBoardCostHigh: 1200,
+  };
+  int get value => values[this] ?? 600;
+}
