+++ lib/pages/home/plan_map/setting_panel/setting_panel_state.dart
@@ -1,11 +1,5 @@
 part of 'setting_panel_cubit.dart';
 
-enum WalkingSpeed { slow, calm, average, prompt, fast }
-enum BikingSpeed { slow, calm, average, prompt, fast }
-enum WalkBoardCost { defaultCost, walkBoardCostHigh }
-enum BikeRentalNetwork { taxi, carSharing, regioRad }
-enum OptimizeType { quick, safe, flat, greenWays, triangle, transfers }
-
 @immutable
 class SettingPanelState extends Equatable {
   static const String _typeWalkingSpeed = "typeWalkingSpeed";
@@ -124,105 +118,3 @@ class SettingPanelState extends Equatable {
         wheelchair,
       ];
 }
-
-
-WalkingSpeed getWalkingSpeed(String key) {
-  return WalkingSpeedExtension.names.keys.firstWhere(
-    (keyE) => keyE.name == key,
-    orElse: () => WalkingSpeed.average,
-  );
-}
-
-BikingSpeed getBikingSpeed(String key) {
-  return BikingSpeedExtension.names.keys.firstWhere(
-    (keyE) => keyE.name == key,
-    orElse: () => BikingSpeed.average,
-  );
-}
-
-BikeRentalNetwork getBikeRentalNetwork(String key) {
-  return BikeRentalNetworkExtension.values.keys.firstWhere(
-    (keyE) => keyE.name == key,
-    orElse: () => BikeRentalNetwork.regioRad,
-  );
-}
-
-extension WalkingSpeedExtension on WalkingSpeed {
-  static const values = <WalkingSpeed, double>{
-    WalkingSpeed.slow: 0.69,
-    WalkingSpeed.calm: 0.97,
-    WalkingSpeed.average: 1.2,
-    WalkingSpeed.prompt: 1.67,
-    WalkingSpeed.fast: 2.22,
-  };
-  static const names = <WalkingSpeed, String>{
-    WalkingSpeed.slow: "slow",
-    WalkingSpeed.calm: "calm",
-    WalkingSpeed.average: "average",
-    WalkingSpeed.prompt: "prompt",
-    WalkingSpeed.fast: "fast",
-  };
-  double get value => values[this] ?? 1.2;
-  String get name => names[this] ?? "slow";
-}
-
-extension BikingSpeedExtension on BikingSpeed {
-  static const values = <BikingSpeed, double>{
-    BikingSpeed.slow: 2.77,
-    BikingSpeed.calm: 4.15,
-    BikingSpeed.average: 5.55,
-    BikingSpeed.prompt: 6.94,
-    BikingSpeed.fast: 8.33,
-  };
-  static const names = <BikingSpeed, String>{
-    BikingSpeed.slow: '10 Km/h',
-    BikingSpeed.calm: '15 Km/h',
-    BikingSpeed.average: '20 Km/h',
-    BikingSpeed.prompt: '25 Km/h',
-    BikingSpeed.fast: '30 Km/h',
-  };
-  double get value => values[this] ?? 5.55;
-  String get name => names[this] ?? "10 Km/h";
-}
-
-extension WalkBoardCostExtension on WalkBoardCost {
-  static const values = <WalkBoardCost, int>{
-    WalkBoardCost.defaultCost: 600,
-    WalkBoardCost.walkBoardCostHigh: 1200,
-  };
-  int get value => values[this] ?? 600;
-}
-
-extension BikeRentalNetworkExtension on BikeRentalNetwork {
-  static const values = <BikeRentalNetwork, String>{
-    BikeRentalNetwork.taxi: 'taxi',
-    BikeRentalNetwork.carSharing: 'car-sharing',
-    BikeRentalNetwork.regioRad: 'regiorad',
-  };
-  String get name => values[this] ?? 'car-sharing';
-}
-
-extension OptimizeTypeExtension on OptimizeType {
-  static const values = <OptimizeType, String>{
-    OptimizeType.quick: 'QUICK',
-    OptimizeType.safe: 'SAFE',
-    OptimizeType.flat: 'FLAT',
-    OptimizeType.greenWays: 'GREENWAYS',
-    OptimizeType.triangle: 'TRIANGLE',
-    OptimizeType.transfers: 'TRANSFERS',
-  };
-  String get name => values[this] ?? 'QUICK';
-}
-
-const defaultTransportModes = <TransportMode>[
-  TransportMode.bus,
-  TransportMode.rail,
-  TransportMode.subway,
-  TransportMode.walk,
-];
-
-const defaultBikeRentalNetworks = <BikeRentalNetwork>[
-  BikeRentalNetwork.taxi,
-  BikeRentalNetwork.carSharing,
-  BikeRentalNetwork.regioRad,
-];
